DynamoDB: Allow TableARN for TableName-parameters
